Rework ModulemdError and ModulemdYamlError
The use of Modulemd[Yaml]ErrorEnum as a type name was incorrect; it
should never have had the Enum suffix (not least because it makes
accessing it longer, but also because g-ir-scanner cannot link it to
the matching GQuark.

This patch makes the ModulemdError and ModulemdYamlError GQuark's
public and creates new typedefs that are properly associated with
them. It does not change the existing enum names so as not to
cause an ABI break in language bindings.

As a bit of good fortune, the language bindings only use the
portion of the identifier that is unique after the prefix, so we can
just use a different prefix for the new enums and not interact
destructively with the old ones.

Signed-off-by: Stephen Gallagher <sgallagh@redhat.com>